backend/requirements.txt
# Neural Network & Image Processing
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2

# Web Framework & API
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP Requests & Scraping
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4

# Image similarity
imagehash==4.3.1
faiss-cpu==1.7.4

